requirements.txt
langchain>=0.1.14
langchain-community>=0.0.30
groq
python-dotenv
streamlit
